pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "olivetti_faces"
version = "0.0.1"
description = "End_to_end_ML_project_olivetti_dataset"
authors = [
  { name = "Andreas_and_Hassan", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = "==3.12.*"

dependencies = [
    "fastapi==0.115.6",
    "torch>=2.9.1",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "numpy>=1.21.6",
    "kagglehub>=0.4.0",
    "invoke>=2.2.0",
    # Added: PCA + SVM
    "scikit-learn>=1.4.0",
    # Added: Hydra config management
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    # Added: experiment tracking
    "wandb>=0.16.0",
    # Added: model persistence
    "joblib>=1.3.0",
    "dvc-gs>=3.0.2",
    "loguru>=0.7.3",
    "evidently>=0.7.20",
    "pandas>=2.3.3",
    "google-cloud-secret-manager>=2.26.0",
    "kfp>=2.15.2",
    "google-cloud-aiplatform>=1.134.0",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mypy==1.19.1",
    "mkdocs-material==9.7.1",
    "mkdocstrings-python==2.0.1",

    # Added: data/model version control
    "dvc>=3.50.0",
]


[tool.uv]
package = true


[tool.setuptools]
package-dir = { "" = "src" }


[tool.setuptools.packages.find]
where = ["src"]


[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = ["tests/*"]

[tool.mypy]
python_version = "3.12.3"
ignore_missing_imports = true
follow_imports = "skip"
show_error_codes = true
pretty = true

[project.scripts]
data = "olivetti_faces.data:app"
train = "olivetti_faces.train:train"
